requirements.txt
django
django-two-factor-auth
django-two-factor-auth[phonenumbers]
django-crispy-forms
phonenumbers
